+++ packages/address/src/loader.ts
@@ -4,7 +4,6 @@ import {
   LoadCountryResponse,
   ResponseError,
   SupportedCountry,
-  SupportedLocale,
 } from './types';
 import query from './graphqlQuery';
 
@@ -19,9 +18,7 @@ const HEADERS = {
   'Access-Control-Allow-Origin': '*',
 };
 
-export async function loadCountries(
-  locale: SupportedLocale,
-): Promise<Country[]> {
+export async function loadCountries(locale: string): Promise<Country[]> {
   const response = await fetch(GRAPHQL_ENDPOINT, {
     method: 'POST',
     headers: HEADERS,
@@ -44,7 +41,7 @@ export async function loadCountries(
 }
 
 export async function loadCountry(
-  locale: SupportedLocale,
+  locale: string,
   countryCode: SupportedCountry,
 ): Promise<Country> {
   const response = await fetch(GRAPHQL_ENDPOINT, {
@@ -74,8 +71,26 @@ class CountryLoaderError extends Error {
   }
 }
 
-function toSupportedLocale(locale: SupportedLocale) {
+const DEFAULT_LOCALE = 'EN';
+const SUPPORTED_LOCALES = [
+  'DA',
+  'DE',
+  'EN',
+  'ES',
+  'FR',
+  'IT',
+  'JA',
+  'NL',
+  'PT',
+  'PT_BR',
+];
+
+function toSupportedLocale(locale: string) {
   const supportedLocale = locale.replace(/-/, '_').toUpperCase();
 
-  return supportedLocale;
+  if (SUPPORTED_LOCALES.includes(supportedLocale)) {
+    return supportedLocale;
+  } else {
+    return DEFAULT_LOCALE;
+  }
 }